requirementsfreeze.txt
absl-py==2.1.0
asttokens==3.0.0
# bresenham3d==1.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.0
crafter==1.8.0
cycler==0.12.1
decorator==4.4.2
dm-control==1.0.9
dm-env==1.6
dm-tree==0.1.8
einops==0.3.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
glfw==2.7.0
grpcio==1.67.1
gym==0.22.0
gym-notices==0.0.8
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.7
labmaze==1.0.6
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.5.0
matplotlib-inline==0.1.7
memory-maze==1.0.3
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.3.5
nbtlib==2.0.4
networkx==3.4.2
noise==1.2.2
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.7.0.72
opensimplex==0.4.5.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==3.20.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
Pygments==2.18.0
PyOpenGL==3.1.1a1
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
requests==2.32.3
ruamel.yaml==0.17.4
scipy==1.14.1
six==1.16.0
snakeviz==2.2.2
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
torch==2.4.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
